- moved shared components to common folder
