+++ src/components/dialogs/index.ts
@@ -1,5 +1,6 @@
 import AccountAuthorizationDialog from './AccountAuthorizationDialog.vue'
 import AddCommentDialog from './AddCommentDialog.vue'
+import CancelPaymentErrorDialog from './CancelPaymentErrorDialog.vue'
 import CoaWarningDialog from './CoaWarningDialog.vue'
 import ConfirmDialog from './ConfirmDialog.vue'
 import DashboardUnavailableDialog from './DashboardUnavailableDialog.vue'
@@ -9,12 +10,12 @@ import LoadCorrectionDialog from './LoadCorrectionDialog.vue'
 import PaymentErrorDialog from './PaymentErrorDialog.vue'
 import ResumeErrorDialog from './ResumeErrorDialog.vue'
 import SaveErrorDialog from './SaveErrorDialog.vue'
-import CancelPaymentErrorDialog from './CancelPaymentErrorDialog.vue'
 import WarningPopover from './WarningPopover.vue'
 
 export {
   AccountAuthorizationDialog,
   AddCommentDialog,
+  CancelPaymentErrorDialog,
   CoaWarningDialog,
   ConfirmDialog,
   DashboardUnavailableDialog,
@@ -24,6 +25,5 @@ export {
   PaymentErrorDialog,
   ResumeErrorDialog,
   SaveErrorDialog,
-  CancelPaymentErrorDialog,
   WarningPopover
 }
