+++ tests/unit/StandaloneDirectorsFiling.spec.ts
@@ -6,9 +6,8 @@ import sinon from 'sinon'
 import { createLocalVue, shallowMount, mount, Wrapper } from '@vue/test-utils'
 import flushPromises from 'flush-promises'
 import CodDate from '@/components/StandaloneDirectorChange/CODDate.vue'
-import Directors from '@/components/AnnualReport/Directors.vue'
-import Certify from '@/components/AnnualReport/Certify.vue'
-import StaffPayment from '@/components/AnnualReport/StaffPayment.vue'
+import Directors from '@/components/common/Directors.vue'
+import { Certify, StaffPayment } from '@/components/common'
 import axios from '@/axios-auth'
 import store from '@/store/store'
 import StandaloneDirectorsFiling from '@/views/StandaloneDirectorsFiling.vue'
